% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph}
\alias{graph}
\title{Graph the relationship between column y and the rest of the x columns}
\usage{
graph(dataset, y)
}
\arguments{
\item{y}{column you want to improve}

\item{data}{raw_dataset}
}
\value{
Returns the result of xy_relation
}
\description{
Graph the relationship between column y and the rest of the x columns
}
