% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_best.R
\name{lm_best}
\alias{lm_best}
\title{Based on the regression model and the given data, the optimal result value and the values of x are returned.}
\usage{
lm_best(dataset, y)
}
\arguments{
\item{y}{column you want to improve}

\item{data}{raw_dataset}
}
\value{
Returns best condition and result
}
\description{
Based on the regression model and the given data, the optimal result value and the values of x are returned.
}
