% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_model.R
\name{lm_model}
\alias{lm_model}
\title{Build multiple linear regression models and perform evaluations}
\usage{
lm_model(dataset, y)
}
\arguments{
\item{y}{column you want to improve}

\item{data}{raw_dataset}
}
\value{
Returns LM Model
}
\description{
Build multiple linear regression models and perform evaluations
}
