% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneshot_model.R
\name{oneshot_model}
\alias{oneshot_model}
\title{After preprocessing the data, building and evaluating the regression and rf models, the function suggests the optimal condition}
\usage{
oneshot_model(raw_dataset, y)
}
\arguments{
\item{y}{column you want to improve}

\item{data}{raw_dataset}
}
\value{
Returns clean data, modeling and best condition and result
}
\description{
After preprocessing the data, building and evaluating the regression and rf models, the function suggests the optimal condition
}
