% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prework1.R
\name{prework1}
\alias{prework1}
\title{Excluding strings of the form of date and id that are unnecessary for analysis, correct missing values with the average value}
\usage{
prework1(raw_dataset)
}
\arguments{
\item{data}{raw_dataset}
}
\value{
Returns the result of preprocessing
}
\description{
Excluding strings of the form of date and id that are unnecessary for analysis, correct missing values with the average value
}
