% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prework2.R
\name{prework2}
\alias{prework2}
\title{Remove outliers and adjust numerous variables to around 10 major variables}
\usage{
prework2(dataset, y)
}
\arguments{
\item{y}{column you want to improve}

\item{data}{raw_dataset}
}
\value{
Returns the result of preprocessing
}
\description{
Remove outliers and adjust numerous variables to around 10 major variables
}
