% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_best.R
\name{rf_best}
\alias{rf_best}
\title{Based on the rf model and the given data, the optimal result value and the values of x are returned.}
\usage{
rf_best(dataset, y)
}
\arguments{
\item{y}{column you want to improve}

\item{data}{raw_dataset}
}
\value{
Returns best condition and result
}
\description{
Based on the rf model and the given data, the optimal result value and the values of x are returned.
}
