% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_model.R
\name{rf_model}
\alias{rf_model}
\title{Build randomForest models and perform evaluations}
\usage{
rf_model(dataset, y)
}
\arguments{
\item{y}{column you want to improve}

\item{data}{raw_dataset}
}
\value{
Returns RF Model
}
\description{
Build randomForest models and perform evaluations
}
